+++ src/sid/policies.py
@@ -2,7 +2,6 @@
 import itertools
 from typing import Any
 from typing import Dict
-from typing import Tuple
 
 import pandas as pd
 from sid.validation import validate_return_is_series_or_ndarray
@@ -10,40 +9,36 @@ from sid.validation import validate_return_is_series_or_ndarray
 
 def apply_contact_policies(
     contact_policies: Dict[str, Dict[str, Any]],
-    recurrent_contacts: pd.DataFrame,
-    random_contacts: pd.DataFrame,
+    contacts: pd.DataFrame,
     states: pd.DataFrame,
     date: pd.Timestamp,
     seed: itertools.count,
-) -> Tuple[pd.DataFrame]:
+) -> pd.DataFrame:
     """Apply policies to contacts."""
-    for contacts in [recurrent_contacts, random_contacts]:
-        if contacts is None:
-            continue
+    for name, policy in contact_policies.items():
+        if policy["start"] <= date <= policy["end"]:
+            func = policy["policy"]
 
-        for name, policy in contact_policies.items():
+            affected_cm = policy.get("affected_contact_model")
+            affected_contacts = (
+                contacts if affected_cm is None else contacts[affected_cm]
+            )
 
-            affected_cm = policy["affected_contact_model"]
-            if affected_cm in contacts.columns:
-                if policy["start"] <= date <= policy["end"]:
-                    func = policy["policy"]
-                    model_specific_contacts = contacts[affected_cm]
+            if isinstance(policy["policy"], (float, int)):
+                affected_contacts = affected_contacts * policy["policy"]
+            else:
+                affected_contacts = func(
+                    states=states,
+                    contacts=affected_contacts,
+                    seed=next(seed),
+                )
 
-                    if isinstance(policy["policy"], (float, int)):
-                        model_specific_contacts = (
-                            model_specific_contacts * policy["policy"]
-                        )
-                    else:
-                        model_specific_contacts = func(
-                            states=states,
-                            contacts=model_specific_contacts,
-                            seed=next(seed),
-                        )
+            if affected_cm is None:
+                contacts = affected_contacts
+            else:
+                affected_contacts = validate_return_is_series_or_ndarray(
+                    affected_contacts, name, "contact_policies", states.index
+                )
+                contacts[affected_cm] = affected_contacts
 
-                    model_specific_contacts = validate_return_is_series_or_ndarray(
-                        model_specific_contacts, name, "contact_policies", states.index
-                    )
-
-                    contacts[affected_cm] = model_specific_contacts
-
-    return recurrent_contacts, random_contacts
+    return contacts